pyproject.toml
[tool.poetry]
name = "create_order"
version = "0.1.0"
description = ""
authors = ["Nawat Ngerncham <67263644+nngerncham@users.noreply.github.com>"]
license = "MIT"
packages = [
    { include = "src", from = "." },
]

[tool.poetry.dependencies]
python = "^3.11"
sqlmodel = "^0.0.12"
python-dotenv = "^1.0.0"
celery = "^5.3.6"
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
opentelemetry-distro = "0.38b0"
opentelemetry-exporter-otlp = "1.17.0"
opentelemetry-instrumentation-celery = "0.38b0"
opentelemetry-sdk = "1.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
